Migrate to Buidler plugin (#1131)
This PR introduces a vanilla setup of buidler development tool. It includes several network configurations both for local development and to interacts with real chains. It also has ganache configured to be used as an alternative local development chain (use with `--network ganache`).

Note the gas reporter packages don't yet support handling the start and stop of the chain. We need to start it manually as explained in https://github.com/cgewecke/buidler-gas-reporter/tree/master#using-buidlerevm-warning.

The changes include:
- Support buidlerevm as default test chain. Depends on https://github.com/aragon/contract-helpers/pull/55.
- Make coverage work with buidler and update to latest solidity-coverage
- Bump to use node 12
- Update `truffle-extract` to support buidler. As discussed in https://github.com/aragon/aragon-apps/pull/1131#discussion_r468681133.